requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-docx
pypdf
pdfplumber
openai
pinecone-client
sqlalchemy
psycopg2-binary
python-dotenv
tqdm
langchain>=0.0.0
requests
pydantic
